+++ internal/domain/experiment.go
@@ -2,6 +2,15 @@ package domain
 
 import "time"
 
+type ExperimentStatus string
+
+const (
+	StatusStarted   ExperimentStatus = "started"
+	StatusUploading ExperimentStatus = "uploading"
+	StatusSuccess   ExperimentStatus = "success"
+	StatusFailed    ExperimentStatus = "failed"
+)
+
 type Experiment struct {
 	ID               string
 	UserID           string
@@ -12,4 +21,6 @@ type Experiment struct {
 	BgrFilePath      string
 	ZipFilePath      string
 	MeteoProfilePath string
+	Status           ExperimentStatus
+	ErrorMessage     string
 }